requirements.txt
Flask
flask-cors
python-dotenv
tensorflow
opencv-python
numpy
google-generativeai